+++ src/db.js
@@ -1,8 +1,14 @@
 import { getDB, closeDBForTest } from './db/connection.js'
+import { CANONICAL_USER_ID, normalizeConversationPartyId } from './db/utils.js'
 
 export { getDB, closeDBForTest }
+export { normalizeConversationPartyId }
+export * from './db/repositories/reminders.js'
+export * from './db/repositories/prefetch.js'
+export * from './db/repositories/media-library.js'
+export * from './db/repositories/audits.js'
+export * from './db/repositories/thread-state.js'
 
-const CANONICAL_USER_ID = 'ID:000001'
 const CANONICAL_AGENT_ENTITY = 'agent:jarvis'
 const CANONICAL_USER_ROOT_MEM_ID = 'person_000001'
 const CANONICAL_AGENT_ROOT_MEM_ID = 'agent_jarvis_identity'
@@ -56,15 +62,6 @@ export function markUISignalsConsumed(ids = []) {
   if (!ids.length) return
   const placeholders = ids.map(() => '?').join(',')
   getDB().prepare(`UPDATE ui_signals SET consumed = 1 WHERE id IN (${placeholders})`).run(...ids)
-}
-
-export function normalizeConversationPartyId(id) {
-  if (!id) return id
-  const text = String(id).trim()
-  if (!text) return text
-  if (/^ID:\d+$/i.test(text)) return `ID:${text.replace(/^ID:/i, '')}`
-  if (/^\d+$/.test(text)) return `ID:${text}`
-  return text
 }
 
 function normalizeMemoryEntity(entity) {
@@ -1341,106 +1338,8 @@ export function getRecentActionLogs(limit = 50, { includeHousekeeping = false, i
   `).all(limit).reverse()
 }
 
-export function createReminder({ userId, dueAt, task, systemMessage, source = '', recurrenceType = null, recurrenceConfig = null }) {
-  const db = getDB()
-  const normalizedUserId = normalizeConversationPartyId(userId || CANONICAL_USER_ID)
-  const configStr = recurrenceConfig ? JSON.stringify(recurrenceConfig) : null
-  return db.prepare(`
-    INSERT INTO reminders (user_id, due_at, task, system_message, status, source, recurrence_type, recurrence_config)
-    VALUES (?, ?, ?, ?, 'pending', ?, ?, ?)
-  `).run(normalizedUserId, dueAt, task, systemMessage, source, recurrenceType, configStr)
-}
-
 // 找到同 user + 同 due_at（精确到分钟）且非周期的待触发提醒，用于合并
-export function findMergeableOneOffReminder(userId, dueAtIsoMinute) {
-  const db = getDB()
-  const normalizedUserId = normalizeConversationPartyId(userId || CANONICAL_USER_ID)
-  return db.prepare(`
-    SELECT * FROM reminders
-    WHERE status = 'pending'
-      AND recurrence_type IS NULL
-      AND user_id = ?
-      AND substr(due_at, 1, 16) = ?
-    ORDER BY id ASC
-    LIMIT 1
-  `).get(normalizedUserId, dueAtIsoMinute) || null
-}
-
-export function appendReminderTask(id, additionalTask, newSystemMessage) {
-  const db = getDB()
-  const row = db.prepare(`SELECT task FROM reminders WHERE id = ?`).get(id)
-  if (!row) return { changes: 0 }
-  const mergedTask = `${row.task}; ${additionalTask}`
-  return db.prepare(`
-    UPDATE reminders
-    SET task = ?, system_message = ?
-    WHERE id = ? AND status = 'pending'
-  `).run(mergedTask, newSystemMessage, id)
-}
-
-export function getDueReminders(now = new Date().toISOString(), limit = 20) {
-  const db = getDB()
-  return db.prepare(`
-    SELECT * FROM reminders
-    WHERE status = 'pending' AND due_at <= ?
-    ORDER BY due_at ASC, id ASC
-    LIMIT ?
-  `).all(now, limit)
-}
-
-export function markReminderFired(id, firedAt = new Date().toISOString()) {
-  const db = getDB()
-  return db.prepare(`
-    UPDATE reminders
-    SET status = 'fired', fired_at = ?
-    WHERE id = ? AND status = 'pending'
-  `).run(firedAt, id)
-}
-
 // 周期提醒触发后：保持 pending，推进 due_at 到下次发生时间
-export function advanceReminderDueAt(id, nextDueAtIso) {
-  const db = getDB()
-  return db.prepare(`
-    UPDATE reminders
-    SET due_at = ?
-    WHERE id = ? AND status = 'pending'
-  `).run(nextDueAtIso, id)
-}
-
-export function cancelReminder(id, cancelledAt = new Date().toISOString()) {
-  const db = getDB()
-  return db.prepare(`
-    UPDATE reminders
-    SET status = 'cancelled', cancelled_at = ?
-    WHERE id = ? AND status = 'pending'
-  `).run(cancelledAt, id)
-}
-
-export function listPendingReminders(limit = 50) {
-  const db = getDB()
-  return db.prepare(`
-    SELECT * FROM reminders
-    WHERE status = 'pending'
-    ORDER BY due_at ASC, id ASC
-    LIMIT ?
-  `).all(limit)
-}
-
-export function getReminderById(id) {
-  const db = getDB()
-  return db.prepare(`SELECT * FROM reminders WHERE id = ?`).get(id) || null
-}
-
-export function getNextPendingReminder() {
-  const db = getDB()
-  return db.prepare(`
-    SELECT * FROM reminders
-    WHERE status = 'pending'
-    ORDER BY due_at ASC, id ASC
-    LIMIT 1
-  `).get() || null
-}
-
 // 按关键词搜索记忆（FTS5 全文搜索，优先相关度排序）
 // 注意：trigram tokenizer 需要查询至少 3 字符；< 3 字符（典型如 2 字中文 ngram）走 LIKE fallback。
 // 软隐藏过滤：FTS5 索引保留全量内容，但 JOIN memories 后用 m.visibility=1 过滤；
@@ -1586,464 +1485,3 @@ export function searchByEmbedding(queryBuffer, limit = 20) {
   scored.sort((a, b) => b._vecScore - a._vecScore)
   return scored.slice(0, Math.max(0, limit))
 }
-
-// ── 预热缓存 ──────────────────────────────────────────────────────────────
-
-export function savePrefetchCache(source, content, ttlMinutes, tags = []) {
-  const db = getDB()
-  const now = new Date()
-  const fetched_at = now.toISOString()
-  const expires_at = new Date(now.getTime() + ttlMinutes * 60 * 1000).toISOString()
-  db.prepare(`
-    INSERT INTO prefetch_cache (source, content, fetched_at, expires_at, tags)
-    VALUES (?, ?, ?, ?, ?)
-    ON CONFLICT(source) DO UPDATE SET
-      content    = excluded.content,
-      fetched_at = excluded.fetched_at,
-      expires_at = excluded.expires_at,
-      tags       = excluded.tags
-  `).run(source, content, fetched_at, expires_at, JSON.stringify(tags))
-}
-
-export function getValidPrefetchCache() {
-  const db = getDB()
-  const now = new Date().toISOString()
-  return db.prepare(`
-    SELECT * FROM prefetch_cache
-    WHERE expires_at > ?
-    ORDER BY fetched_at DESC
-  `).all(now)
-}
-
-export function clearExpiredPrefetchCache() {
-  const db = getDB()
-  const now = new Date().toISOString()
-  db.prepare(`DELETE FROM prefetch_cache WHERE expires_at <= ?`).run(now)
-}
-
-// ── 预热任务管理 ──────────────────────────────────────────────────────────
-
-export function upsertPrefetchTask({ source, label, url, ttlMinutes = 60, tags = [] }) {
-  const db = getDB()
-  const now = new Date().toISOString()
-  db.prepare(`
-    INSERT INTO prefetch_tasks (source, label, url, ttl_minutes, tags, enabled, updated_at)
-    VALUES (?, ?, ?, ?, ?, 1, ?)
-    ON CONFLICT(source) DO UPDATE SET
-      label       = excluded.label,
-      url         = excluded.url,
-      ttl_minutes = excluded.ttl_minutes,
-      tags        = excluded.tags,
-      enabled     = 1,
-      updated_at  = excluded.updated_at
-  `).run(source, label, url, ttlMinutes, JSON.stringify(tags), now)
-}
-
-export function removePrefetchTask(source) {
-  const db = getDB()
-  const result = db.prepare(`DELETE FROM prefetch_tasks WHERE source = ?`).run(source)
-  return result.changes > 0
-}
-
-export function listPrefetchTasks() {
-  const db = getDB()
-  return db.prepare(`SELECT * FROM prefetch_tasks ORDER BY created_at ASC`).all()
-}
-
-export function getEnabledPrefetchTasks() {
-  const db = getDB()
-  return db.prepare(`SELECT * FROM prefetch_tasks WHERE enabled = 1 ORDER BY created_at ASC`).all()
-}
-
-// ── 媒体播放历史 ──────────────────────────────────────────────────────────────
-
-export function upsertMediaHistory({ kind, url, title = '', videoId = null, platform = null }) {
-  const db = getDB()
-  const now = new Date().toISOString()
-  db.prepare(`
-    INSERT INTO media_history (kind, url, title, video_id, platform, played_at)
-    VALUES (?, ?, ?, ?, ?, ?)
-    ON CONFLICT(url) DO UPDATE SET
-      title     = excluded.title,
-      played_at = excluded.played_at
-  `).run(kind, url, title, videoId || null, platform || null, now)
-}
-
-export function getMediaHistory(limit = 30) {
-  const db = getDB()
-  return db.prepare(`
-    SELECT * FROM media_history ORDER BY played_at DESC LIMIT ?
-  `).all(limit)
-}
-
-// ── Music Library ────────────────────────────────────────────────────────────
-
-export function upsertMusicTrack({ title = '', artist = '', album = '', filePath, duration = 0, lrc = '', cover = '', sourceUrl = '' }) {
-  const db = getDB()
-  db.prepare(`
-    INSERT INTO music_library (title, artist, album, file_path, duration, lrc, cover, source_url)
-    VALUES (?, ?, ?, ?, ?, ?, ?, ?)
-    ON CONFLICT(file_path) DO UPDATE SET
-      title      = excluded.title,
-      artist     = excluded.artist,
-      album      = excluded.album,
-      duration   = excluded.duration,
-      lrc        = CASE WHEN excluded.lrc != '' THEN excluded.lrc ELSE lrc END,
-      cover      = CASE WHEN excluded.cover != '' THEN excluded.cover ELSE cover END,
-      source_url = CASE WHEN excluded.source_url != '' THEN excluded.source_url ELSE source_url END
-  `).run(title, artist, album, filePath, duration, lrc, cover, sourceUrl)
-  return db.prepare(`SELECT * FROM music_library WHERE file_path = ?`).get(filePath)
-}
-
-export function getMusicTrack(id) {
-  return getDB().prepare(`SELECT * FROM music_library WHERE id = ?`).get(id)
-}
-
-export function searchMusicLibrary(query, limit = 20) {
-  const db = getDB()
-  const tokens = String(query || '').trim().split(/\s+/).filter(Boolean)
-  if (!tokens.length) return []
-  // 每个词都要命中 title/artist/album 之一（per-token OR，token 之间 AND）：
-  // 这样 "歌名 歌手" 能匹配到 title 只含歌名、artist 只含歌手的行，避免漏命中后白跑下载。
-  const clauses = tokens.map(() => '(title LIKE ? OR artist LIKE ? OR album LIKE ?)')
-  const params = []
-  for (const t of tokens) { const like = `%${t}%`; params.push(like, like, like) }
-  return db.prepare(`
-    SELECT * FROM music_library
-    WHERE ${clauses.join(' AND ')}
-    ORDER BY added_at DESC LIMIT ?
-  `).all(...params, limit)
-}
-
-export function listMusicLibrary(limit = 50) {
-  return getDB().prepare(`SELECT * FROM music_library ORDER BY added_at DESC LIMIT ?`).all(limit)
-}
-
-export function updateMusicLrc(id, lrc) {
-  getDB().prepare(`UPDATE music_library SET lrc = ? WHERE id = ?`).run(lrc, id)
-}
-
-export function deleteMusicTrack(id) {
-  getDB().prepare(`DELETE FROM music_library WHERE id = ?`).run(id)
-}
-
-// ============================================================
-// focus_stack —— 动态上下文记忆池 5c 步：注意力焦点栈持久化
-// ============================================================
-//
-// loadFocusStack: 启动时一次性读出整栈（按 depth ASC）；任何异常都返回 []，
-//   不阻塞主流程。frame 形状与内存中的 state.focusStack[i] 完全一致：
-//   { topic, startedAt, startedAtTick, lastSeenTick, hitCount, conclusions }
-//
-// saveFocusStack: 整栈原子替换。先 DELETE 再 INSERT，全部包在 transaction 里。
-//   focus.js 只在内存里改 state.focusStack，所以 index.js 在每次 updateFocusFrame
-//   返回非 noop 时主动调；focus-compress.js 也通过 onConclusionAttached 回调触发。
-//   写库失败 console.warn 后吞掉——专注栈丢一次远比阻塞主对话轻。
-const FOCUS_STACK_RESTORE_TTL_MS = 24 * 60 * 60 * 1000
-
-export function loadFocusStack() {
-  const db = getDB()
-  try {
-    const rows = db.prepare(`SELECT * FROM focus_stack ORDER BY depth ASC`).all()
-    if (rows.length > 0) {
-      const newest = rows
-        .map(r => Date.parse(r.updated_at || r.started_at || ''))
-        .filter(Number.isFinite)
-        .sort((a, b) => b - a)[0]
-      if (Number.isFinite(newest) && Date.now() - newest > FOCUS_STACK_RESTORE_TTL_MS) {
-        db.prepare(`DELETE FROM focus_stack`).run()
-        console.log('[focus-persist] dropped stale persisted focus_stack on startup')
-        return []
-      }
-    }
-    return rows.map(r => ({
-      topic: JSON.parse(r.topic || '[]'),
-      startedAt: r.started_at,
-      startedAtTick: r.started_at_tick,
-      lastSeenTick: r.last_seen_tick,
-      hitCount: r.hit_count,
-      conclusions: JSON.parse(r.conclusions || '[]'),
-    }))
-  } catch {
-    return []
-  }
-}
-
-export function saveFocusStack(stack) {
-  const db = getDB()
-  try {
-    const tx = db.transaction((frames) => {
-      db.prepare(`DELETE FROM focus_stack`).run()
-      const insert = db.prepare(`
-        INSERT INTO focus_stack (depth, topic, started_at, started_at_tick, last_seen_tick, hit_count, conclusions)
-        VALUES (?, ?, ?, ?, ?, ?, ?)
-      `)
-      for (let i = 0; i < frames.length; i++) {
-        const f = frames[i]
-        insert.run(
-          i,
-          JSON.stringify(f.topic || []),
-          f.startedAt || new Date().toISOString(),
-          f.startedAtTick || 0,
-          f.lastSeenTick || 0,
-          f.hitCount || 1,
-          JSON.stringify(f.conclusions || [])
-        )
-      }
-    })
-    tx(stack || [])
-  } catch (err) {
-    console.warn('[focus-persist] saveFocusStack failed:', err.message)
-  }
-}
-
-// ============================================================
-// threads / commitments —— 线索模型（DynamicMemoryPool.md 第 8 章）持久化
-// ============================================================
-//
-// loadThreadState: 启动时一次性读出 { threads, foregroundId, commitments }。
-//   只加载 status='open' 的承诺所属线索 + 最近 7 天活跃线索（冷线索留在仓库，召回另走记忆）。
-//   threads 为空且 focus_stack 有货 → 返回 null 让上层走 migrateFocusStackToThreads。
-// saveThreadState: 整态原子替换（transaction）。与 saveFocusStack 同款的"丢一次不如阻塞"取舍。
-const THREAD_LOAD_WINDOW_MS = 7 * 24 * 60 * 60 * 1000
-
-function rowToThread(r) {
-  const topic = JSON.parse(r.topic || '[]')
-  const signature = JSON.parse(r.signature || '[]')
-  return {
-    id: r.id,
-    topic,
-    signature: signature.length > 0 ? signature : [...topic],
-    label: r.label || '',
-    summary: r.summary || '',
-    conclusions: JSON.parse(r.conclusions || '[]'),
-    status: r.status || 'open',
-    createdAt: r.created_at,
-    lastEventAt: r.last_event_at,
-    lastEventTick: r.last_event_tick || 0,
-    hitCount: r.hit_count || 1,
-    lastSummaryAt: r.last_summary_at || '',
-  }
-}
-
-export function loadThreadState() {
-  const db = getDB()
-  try {
-    const threadRows = db.prepare(`SELECT * FROM threads`).all()
-    if (threadRows.length === 0) return null
-    const commitmentRows = db.prepare(`SELECT * FROM commitments WHERE status = 'open'`).all()
-    const openThreadIds = new Set(commitmentRows.map(r => r.thread_id))
-    const cutoff = Date.now() - THREAD_LOAD_WINDOW_MS
-    const threads = threadRows
-      .filter(r => openThreadIds.has(r.id) || (Date.parse(r.last_event_at || '') || 0) >= cutoff)
-      .map(rowToThread)
-    const fgRow = db.prepare(`SELECT value FROM thread_state WHERE key = 'foregroundId'`).get()
-    let foregroundId = fgRow?.value || null
-    if (foregroundId && !threads.some(t => t.id === foregroundId)) foregroundId = null
-    const commitments = commitmentRows.map(r => ({
-      id: r.id,
-      threadId: r.thread_id,
-      text: r.text || '',
-      status: r.status || 'open',
-      channel: r.channel || '',
-      createdAt: r.created_at,
-      closedAt: r.closed_at || null,
-    }))
-    return { threads, foregroundId, commitments }
-  } catch {
-    return null
-  }
-}
-
-export function saveThreadState(threadState) {
-  const db = getDB()
-  const ts = threadState || {}
-  try {
-    const tx = db.transaction(() => {
-      const upsertThread = db.prepare(`
-        INSERT INTO threads (id, topic, signature, label, summary, conclusions, status, created_at, last_event_at, last_event_tick, hit_count, last_summary_at, updated_at)
-        VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, datetime('now'))
-        ON CONFLICT(id) DO UPDATE SET
-          topic = excluded.topic, signature = excluded.signature, label = excluded.label, summary = excluded.summary,
-          conclusions = excluded.conclusions, status = excluded.status,
-          last_event_at = excluded.last_event_at, last_event_tick = excluded.last_event_tick,
-          hit_count = excluded.hit_count, last_summary_at = excluded.last_summary_at,
-          updated_at = datetime('now')
-      `)
-      // 注意：threads 表不 DELETE——线索数据只增不删（被 evict 出内存的线索仍在仓库）。
-      for (const t of (ts.threads || [])) {
-        upsertThread.run(
-          t.id,
-          JSON.stringify(t.topic || []),
-          JSON.stringify(t.signature || t.topic || []),
-          t.label || '',
-          t.summary || '',
-          JSON.stringify(t.conclusions || []),
-          t.status || 'open',
-          t.createdAt || new Date().toISOString(),
-          t.lastEventAt || new Date().toISOString(),
-          t.lastEventTick || 0,
-          t.hitCount || 1,
-          t.lastSummaryAt || ''
-        )
-      }
-      const upsertCommitment = db.prepare(`
-        INSERT INTO commitments (id, thread_id, text, status, channel, created_at, closed_at)
-        VALUES (?, ?, ?, ?, ?, ?, ?)
-        ON CONFLICT(id) DO UPDATE SET
-          thread_id = excluded.thread_id, text = excluded.text, status = excluded.status,
-          channel = excluded.channel, closed_at = excluded.closed_at
-      `)
-      for (const c of (ts.commitments || [])) {
-        upsertCommitment.run(
-          c.id, c.threadId, c.text || '', c.status || 'open',
-          c.channel || '', c.createdAt || new Date().toISOString(), c.closedAt || null
-        )
-      }
-      db.prepare(`
-        INSERT INTO thread_state (key, value) VALUES ('foregroundId', ?)
-        ON CONFLICT(key) DO UPDATE SET value = excluded.value
-      `).run(ts.foregroundId || '')
-      // 内存中被 merge 掉的线索：仓库行标记 closed（不删行）
-      if (Array.isArray(ts.mergedAwayIds)) {
-        const close = db.prepare(`UPDATE threads SET status = 'merged', updated_at = datetime('now') WHERE id = ?`)
-        for (const id of ts.mergedAwayIds) close.run(id)
-      }
-    })
-    tx()
-  } catch (err) {
-    console.warn('[thread-persist] saveThreadState failed:', err.message)
-  }
-}
-
-// ───────── 记忆观测层（Memory-Optimization v0.1 Phase 0） ─────────
-// 所有 audit 写入都是 best-effort：失败只记 warn，不抛出，不影响主流程。
-
-export function insertRecallAudit({
-  turn_label = null,
-  from_id = null,
-  channel = null,
-  query_text = '',
-  matched_mem_ids = [],
-  chosen_count = 0,
-  event_type_dist = {},
-  latency_ms = null,
-  source = null,
-} = {}) {
-  try {
-    const ids = Array.isArray(matched_mem_ids) ? matched_mem_ids : []
-    getDB().prepare(`
-      INSERT INTO recall_audit (
-        turn_label, from_id, channel, query_text,
-        matched_mem_ids, matched_count, chosen_count,
-        event_type_dist, latency_ms, source
-      ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-    `).run(
-      turn_label,
-      from_id,
-      channel,
-      (query_text || '').slice(0, 1000),
-      JSON.stringify(ids),
-      ids.length,
-      chosen_count,
-      JSON.stringify(event_type_dist || {}),
-      latency_ms,
-      source
-    )
-  } catch (err) {
-    console.warn('[recall_audit] insert failed:', err.message)
-  }
-}
-
-export function insertExtractAudit({
-  turn_label = null,
-  from_id = null,
-  channel = null,
-  turn_summary = '',
-  extracted_mem_ids = [],
-  event_type_dist = {},
-  latency_ms = null,
-  skipped = false,
-  skip_reason = null,
-} = {}) {
-  try {
-    const ids = Array.isArray(extracted_mem_ids) ? extracted_mem_ids : []
-    getDB().prepare(`
-      INSERT INTO extract_audit (
-        turn_label, from_id, channel, turn_summary,
-        extracted_mem_ids, extracted_count,
-        event_type_dist, latency_ms, skipped, skip_reason
-      ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-    `).run(
-      turn_label,
-      from_id,
-      channel,
-      (turn_summary || '').slice(0, 500),
-      JSON.stringify(ids),
-      ids.length,
-      JSON.stringify(event_type_dist || {}),
-      latency_ms,
-      skipped ? 1 : 0,
-      skip_reason
-    )
-  } catch (err) {
-    console.warn('[extract_audit] insert failed:', err.message)
-  }
-}
-
-export function getRecentRecallAudits(limit = 50) {
-  try {
-    return getDB().prepare(`SELECT * FROM recall_audit ORDER BY id DESC LIMIT ?`).all(limit)
-  } catch (err) {
-    console.warn('[recall_audit] read failed:', err.message)
-    return []
-  }
-}
-
-export function getRecentExtractAudits(limit = 50) {
-  try {
-    return getDB().prepare(`SELECT * FROM extract_audit ORDER BY id DESC LIMIT ?`).all(limit)
-  } catch (err) {
-    console.warn('[extract_audit] read failed:', err.message)
-    return []
-  }
-}
-
-export function getRecallAuditStats({ sinceIso = null } = {}) {
-  try {
-    const sinceClause = sinceIso ? 'WHERE created_at >= ?' : ''
-    const args = sinceIso ? [sinceIso] : []
-    return getDB().prepare(`
-      SELECT
-        COUNT(*) AS total,
-        AVG(matched_count) AS avg_matched,
-        AVG(chosen_count)  AS avg_chosen,
-        AVG(latency_ms)    AS avg_latency_ms,
-        MAX(latency_ms)    AS max_latency_ms,
-        SUM(CASE WHEN matched_count = 0 THEN 1 ELSE 0 END) AS zero_match_count
-      FROM recall_audit ${sinceClause}
-    `).get(...args)
-  } catch (err) {
-    console.warn('[recall_audit] stats failed:', err.message)
-    return null
-  }
-}
-
-export function getExtractAuditStats({ sinceIso = null } = {}) {
-  try {
-    const sinceClause = sinceIso ? 'WHERE created_at >= ?' : ''
-    const args = sinceIso ? [sinceIso] : []
-    return getDB().prepare(`
-      SELECT
-        COUNT(*)            AS total,
-        AVG(extracted_count) AS avg_extracted,
-        AVG(latency_ms)      AS avg_latency_ms,
-        SUM(skipped)         AS skipped_count
-      FROM extract_audit ${sinceClause}
-    `).get(...args)
-  } catch (err) {
-    console.warn('[extract_audit] stats failed:', err.message)
-    return null
-  }
-}
-
