+++ src/test-db-repository-helper.js
@@ -1,0 +1,28 @@
+import fs from 'node:fs'
+import os from 'node:os'
+import path from 'node:path'
+
+export async function runRepositoryTest(name, test) {
+  const tempRoot = fs.mkdtempSync(path.join(os.tmpdir(), `bailongma-${name}-`))
+  const resolvedTempRoot = path.resolve(tempRoot)
+  const resolvedOsTemp = path.resolve(os.tmpdir())
+
+  if (!resolvedTempRoot.startsWith(`${resolvedOsTemp}${path.sep}`)) {
+    throw new Error(`Refusing to use unsafe temp path: ${resolvedTempRoot}`)
+  }
+
+  process.env.BAILONGMA_USER_DIR = resolvedTempRoot
+  process.env.BAILONGMA_RESOURCES_DIR = path.resolve('.')
+
+  try {
+    const connection = await import('./db/connection.js')
+    await test(connection)
+    connection.closeDBForTest()
+  } finally {
+    const { closeDBForTest } = await import('./db/connection.js')
+    closeDBForTest()
+    fs.rmSync(resolvedTempRoot, { recursive: true, force: true })
+  }
+
+  console.log(`${name} repository tests passed`)
+}
